simplify after merging
